+++ tests/AssertTest.php
@@ -70,9 +70,13 @@ final class AssertTest extends TestCase
         }, "abc");
         $this->assertMatchesRegExp('/abc/', "1abc2");
         $this->assertMatchesFile(__DIR__ . "/test.txt", "abc\n");
-        $this->assertArrayOfClass(
-            \stdClass::class,
+        $this->assertArrayOfType(
+            stdClass::class,
             [new stdClass(), new stdClass(), ]
+        );
+        $this->assertArrayOfType(
+            "float",
+            [1.5, 2.0, ]
         );
         $this->assertTriggersDeprecation(function () {
             trigger_error("test", E_USER_DEPRECATED);
@@ -222,81 +226,86 @@ final class AssertTest extends TestCase
             $this->assertMatchesRegExp('/abc/', "def");
         }, AssertionFailedException::class, "Test 57 failed. The string does not match regular expression.");
         $this->assertThrowsException(function () {
-            $this->assertArrayOfClass(
-                \stdClass::class,
-                [new stdClass(), new DummyEngine(), "abc", ] // @phpstan-ignore argument.type
-            );
-        }, AssertionFailedException::class, "Test 59 failed. The array does not contain only instances of stdClass.");
-        $this->assertThrowsException(function () {
-            $this->assertMatchesFile(__DIR__ . "/non_existing.txt", "");
-        }, AssertionFailedException::class,
-            "Test 61 failed. File " . __DIR__ . "/non_existing.txt could not be loaded.");
-        $this->assertThrowsException(function () {
-            $this->assertMatchesFile(__DIR__ . "/test.txt", "");
-        }, AssertionFailedException::class, "Test 63 failed. The value is not 'abc\n' but ''.");
-        $this->assertThrowsException(function () {
-            $this->assertArrayOfClass(
-                \stdClass::class,
+            $this->assertArrayOfType(
+                stdClass::class,
                 []
             );
-        }, AssertionFailedException::class, "Test 65 failed. The array is empty.");
+        }, AssertionFailedException::class, "Test 59 failed. The array is empty.");
+        $this->assertThrowsException(function () {
+            $this->assertArrayOfType(
+                "string",
+                ["abc", 123, ]
+            );
+        }, AssertionFailedException::class, "Test 61 failed. The array does not contain only values of type string.");
+        $this->assertThrowsException(function () {
+            $this->assertArrayOfType(
+                stdClass::class,
+                [new stdClass(), new DummyEngine(), ]
+            );
+        }, AssertionFailedException::class, "Test 63 failed. The array does not contain only values of type string.");
+        $this->assertThrowsException(function () {
+            $this->assertMatchesFile(__DIR__ . "/non_existing.txt", "");
+        }, AssertionFailedException::class, "Test 65 failed. File " . __DIR__ . "/non_existing.txt could not be loaded.");
+        $this->assertThrowsException(function () {
+            $this->assertMatchesFile(__DIR__ . "/test.txt", "");
+        }, AssertionFailedException::class, "Test 67 failed. The value is not 'abc\n' but ''.");
         $this->assertThrowsException(function () {
             $this->assertTriggersDeprecation(function () {
             });
-        }, AssertionFailedException::class, "Test 67 failed. Expected a deprecation but none was triggered.");
+        }, AssertionFailedException::class, "Test 69 failed. Expected a deprecation but none was triggered.");
         $this->assertThrowsException(function () {
             $this->assertTriggersDeprecation(function () {
                 trigger_error("test", E_USER_DEPRECATED);
             }, "abc");
-        }, AssertionFailedException::class, "Test 69 failed. Expected deprecation 'abc' but 'test' was triggered.");
+        }, AssertionFailedException::class, "Test 71 failed. Expected deprecation 'abc' but 'test' was triggered.");
         $this->assertThrowsException(function () {
             $this->assertTriggersNoDeprecation(function () {
                 trigger_error("test", E_USER_DEPRECATED);
             });
-        }, AssertionFailedException::class, "Test 71 failed. Expected no deprecation but 'test' was triggered.");
+        }, AssertionFailedException::class, "Test 73 failed. Expected no deprecation but 'test' was triggered.");
         $this->assertThrowsException(function () {
             $this->assertArrayHasKey("test", ["abc" => 1, "def" => 2, ]);
-        }, AssertionFailedException::class, "Test 73 failed. The array does not contain key 'test'.");
+        }, AssertionFailedException::class, "Test 75 failed. The array does not contain key 'test'.");
         $this->assertThrowsException(function () {
             $this->assertArrayHasKey(5, [0, 5, 10, ]);
-        }, AssertionFailedException::class, "Test 75 failed. The array does not contain key 5.");
+        }, AssertionFailedException::class, "Test 77 failed. The array does not contain key 5.");
         $this->assertThrowsException(function () {
             $this->assertArrayHasKey("test", new ArrayObject());
-        }, AssertionFailedException::class, "Test 77 failed. The array does not contain key 'test'.");
+        }, AssertionFailedException::class, "Test 79 failed. The array does not contain key 'test'.");
         $this->assertThrowsException(function () {
             $this->assertArrayNotHasKey("abc", ["abc" => 1, "def" => 2, ]);
-        }, AssertionFailedException::class, "Test 79 failed. The array contains key 'abc'.");
+        }, AssertionFailedException::class, "Test 81 failed. The array contains key 'abc'.");
         $this->assertThrowsException(function () {
             $this->assertArrayNotHasKey(1, [0, 5, 10, ]);
-        }, AssertionFailedException::class, "Test 81 failed. The array contains key 1.");
+        }, AssertionFailedException::class, "Test 83 failed. The array contains key 1.");
         $this->assertThrowsException(function () {
             $arrayObject = new ArrayObject();
             $arrayObject->offsetSet("test", "abc");
             $this->assertArrayNotHasKey("test", $arrayObject);
-        }, AssertionFailedException::class, "Test 83 failed. The array contains key 'test'.");
+        }, AssertionFailedException::class, "Test 85 failed. The array contains key 'test'.");
         $this->assertThrowsException(function () {
             $this->assertSameSize([], [0, ]);
-        }, AssertionFailedException::class, "Test 85 failed. Actual count is 1 not 0.");
+        }, AssertionFailedException::class, "Test 87 failed. Actual count is 1 not 0.");
         $this->assertThrowsException(function () {
             $arrayObject = new ArrayObject();
             $arrayObject->offsetSet("test", "abc");
             $this->assertSameSize($arrayObject, []);
-        }, AssertionFailedException::class, "Test 87 failed. Actual count is 0 not 1.");
+        }, AssertionFailedException::class, "Test 89 failed. Actual count is 0 not 1.");
         $this->assertThrowsException(function () {
             $this->assertSameSize(new ArrayObject(), [0, ]);
-        }, AssertionFailedException::class, "Test 89 failed. Actual count is 1 not 0.");
+        }, AssertionFailedException::class, "Test 91 failed. Actual count is 1 not 0.");
         $this->assertThrowsException(function () {
             $this->assertFileExists("/path/to/non/existing/file");
-        }, AssertionFailedException::class, "Test 91 failed. File /path/to/non/existing/file does not exist.");
+        }, AssertionFailedException::class, "Test 93 failed. File /path/to/non/existing/file does not exist.");
         $this->assertThrowsException(function () {
             $this->assertFileNotExists(__FILE__);
-        }, AssertionFailedException::class, "Test 93 failed. File " . __FILE__ . " exists.");
+        }, AssertionFailedException::class, "Test 95 failed. File " . __FILE__ . " exists.");
         $this->assertThrowsException(function () {
             $this->assertDirectoryExists("/path/to/non/existing/dir");
-        }, AssertionFailedException::class, "Test 95 failed. Directory /path/to/non/existing/dir does not exist.");
+        }, AssertionFailedException::class, "Test 97 failed. Directory /path/to/non/existing/dir does not exist.");
         $this->assertThrowsException(function () {
             $this->assertDirectoryNotExists(__DIR__);
-        }, AssertionFailedException::class, "Test 97 failed. Directory " . __DIR__ . " exists.");
+        }, AssertionFailedException::class, "Test 99 failed. Directory " . __DIR__ . " exists.");
     }
 
     /**